Refactor test suite structure and update namespaces
